+++ lib/cuprum/rails/records.rb
@@ -6,5 +6,6 @@ module Cuprum::Rails
   # The Records collection wraps ActiveRecord model persistence and querying.
   module Records
     autoload :Collection, 'cuprum/rails/records/collection'
+    autoload :Repository, 'cuprum/rails/records/repository'
   end
 end