+++ spec/cuprum/rails/actions/destroy_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe Cuprum::Rails::Actions::Destroy do
 
   subject(:action) { described_class.new }
 
-  let(:repository) { Cuprum::Rails::Repository.new }
+  let(:repository) { Cuprum::Rails::Records::Repository.new }
   let(:resource) do
     Cuprum::Rails::Resource.new(entity_class: Book)
   end