+++ spec/cuprum/rails/actions/middleware/resources/query_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 require 'cuprum/rails/actions/middleware/resources/query'
+require 'cuprum/rails/records/repository'
 require 'cuprum/rails/request'
-require 'cuprum/rails/repository'
 require 'cuprum/rails/resource'
 
 RSpec.describe Cuprum::Rails::Actions::Middleware::Resources::Query do
@@ -36,7 +36,7 @@ RSpec.describe Cuprum::Rails::Actions::Middleware::Resources::Query do
     let(:next_result)  { Cuprum::Result.new(value: { 'ok' => true }) }
     let(:next_command) { instance_double(Cuprum::Command, call: next_result) }
     let(:request)      { Cuprum::Rails::Request.new }
-    let(:repository)   { Cuprum::Rails::Repository.new }
+    let(:repository)   { Cuprum::Rails::Records::Repository.new }
     let(:resource)     { Cuprum::Rails::Resource.new(name: 'authors') }
     let(:options)      { {} }