+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 Refactored the existing collection to `Cuprum::Rails::Records::Collection`.
 
 - Refactor `Collection` to `Records::Collection`. Creating a `Cuprum::Rails::Collection` will print a deprecation warning.
+- Refactor `Repository` to `Records::Repository`. Creating a `Cuprum::Rails::Repository` will print a deprecation warning.
 
 ## O.2.0